+++ sendhooks/queue/worker.go
@@ -8,13 +8,10 @@ we also handle the retries and the exponential backoff logic for sending the hoo
 import (
 	"context"
 	"fmt"
+	"sendhooks/adapter"
 	"sendhooks/logging"
 	"sendhooks/sender"
 	"time"
-
-	"github.com/go-redis/redis/v8"
-
-	redisClient "sendhooks/redis"
 )
 
 const (
@@ -26,17 +23,18 @@ const (
 	maxBackoff     time.Duration = time.Hour
 )
 
-func ProcessWebhooks(ctx context.Context, webhookQueue chan redisClient.WebhookPayload, client *redis.Client, configuration redisClient.Configuration) {
+func ProcessWebhooks(ctx context.Context, webhookQueue chan adapter.WebhookPayload, configuration adapter.Configuration, queueAdapter adapter.Adapter) {
 
 	for payload := range webhookQueue {
-		go sendWebhookWithRetries(ctx, payload, client, configuration)
+		go sendWebhookWithRetries(ctx, payload, configuration, queueAdapter)
 	}
 }
 
-func sendWebhookWithRetries(ctx context.Context, payload redisClient.WebhookPayload, client *redis.Client, configuration redisClient.Configuration) {
-	if err, created := retryWithExponentialBackoff(ctx, payload, client, configuration); err != nil {
+func sendWebhookWithRetries(ctx context.Context, payload adapter.WebhookPayload, configuration adapter.Configuration, queueAdapter adapter.Adapter) {
+
+	if err, created := retryWithExponentialBackoff(ctx, payload, configuration, queueAdapter); err != nil {
 		logging.WebhookLogger(logging.WarningType, fmt.Errorf("failed to send sendhooks after maximum retries. WebhookID : %s", payload.WebhookID))
-		err := redisClient.PublishStatus(ctx, payload.WebhookID, payload.URL, created, "", "failed", err.Error(), client, configuration)
+		err := queueAdapter.PublishStatus(ctx, payload.WebhookID, payload.URL, created, "", "failed", err.Error())
 		if err != nil {
 			logging.WebhookLogger(logging.WarningType, fmt.Errorf("error publishing status update: WebhookID : %s ", payload.WebhookID))
 		}
@@ -54,7 +52,7 @@ func calculateBackoff(currentBackoff time.Duration) time.Duration {
 	return nextBackoff
 }
 
-func retryWithExponentialBackoff(context context.Context, payload redisClient.WebhookPayload, client *redis.Client, configuration redisClient.Configuration) (error, string) {
+func retryWithExponentialBackoff(context context.Context, payload adapter.WebhookPayload, configuration adapter.Configuration, queueAdapter adapter.Adapter) (error, string) {
 	retries := 0
 	backoffTime := initialBackoff
 	var requestError error
@@ -90,7 +88,7 @@ func retryWithExponentialBackoff(context context.Context, payload redisClient.We
 
 	delivered := time.Now().String()
 
-	err := redisClient.PublishStatus(context, payload.WebhookID, payload.URL, created, delivered, "success", "", client, configuration)
+	err := queueAdapter.PublishStatus(context, payload.WebhookID, payload.URL, created, delivered, "success", "")
 	if err != nil {
 		logging.WebhookLogger(logging.WarningType, fmt.Errorf("error publishing status update: WebhookID : %s ", payload.WebhookID))
 	}